requirements.txt
runpod
diffusers
transformers
accelerate
safetensors
imageio
imageio-ffmpeg
opencv-python-headless
pillow
huggingface_hub
torch
